Templatize QueueForOneReaderOneIrqWriter for trivial types (#12)
* Convert QueueForOneReaderOneIrqWriter to template class

* Remove QueueForOneReaderOneIrqWriter.cpp from cmakelists as well

* QueueForOneReaderOneIrqWriterTest: Assert on trivial type

* QueueForOneReaderOneIrqWriter: Address review comments
